OnMessage Event relocated
