+++ src/modules/peach.py
@@ -14,7 +14,7 @@ class Peach:
 
     @commands.command(name="replaceass", aliases=["peach"], hidden=True)
     async def replace_ass(self, ctx):    # Sends the ass substitution message
-        with io.open("data/user_data.json") as f:
+        with io.open("../data/user_data.json") as f:
             lmao_count_data = json.load(f)
             try:
                 lmao_count_data[str(ctx.author.id)]["lmao_count"] += 1
@@ -22,7 +22,7 @@ class Peach:
                 lmao_count_data[str(ctx.author.id)] = {"lmao_count": 1}
             lmao_count_data[str(ctx.author.id)]["username"] = str(ctx.author)
             new_user_data = json.dumps(lmao_count_data, indent=4)
-            with io.open("data/user_data.json", "w+", encoding="utf-8") as fo:
+            with io.open("../data/user_data.json", "w+", encoding="utf-8") as fo:
                 fo.write(new_user_data)
         x = random.randint(1, 100)
         if ctx.guild is None or x <= lbvars.get_react_chance(ctx.guild.id):
